Add files via upload
Portfolio 1:
This Power BI file intends to showcase data in a structured way to provide comprehensive insights through visualizations. Each of the dashboards focus on a particular aspect of business. The organization and approach of each of the visuals aim to deliver clear and actionable insights to stakeholders at various levels. 
The dataset is that of Adventure Works (taken online). The file contains 4 dashboards - Executive Dashboard, Geographical Map, Product Details and Customer Details. The dashboards are dynamic and built to have the capabilities to provide a wholistic view of the business performance while also accommodating the facility to drill down information of individual elements e.g. a product, a region or that of a customer. 
Additionally, the dashboards have built-in buttons to help navigate between each other.

The following is a short description of each of the dashboards:
Executive Dashboard - This aims to serve as an executive summary, providing high-level insights and key performance indicators (KPIs) critical for decision-makers.
The display exhibits a comprehensive overview of the data and presents a possibility to have interactive elements to drill down into specific Product metrics. The monthly revenue, profits, etc. are visually placed to capture attention. Different colours are used to highlight the health and performance of the metrics. The thematic grey represents a healthy business while the red markings indicate the concerning areas. 

Map Dashboard - The visualizations include regions of Europe, North America and Pacific and illustrates regional business performance and distributions.

Product Details Dashboard - This provides detailed information of the business performance about all products. It also provides product-specific analytics and their performance related to targets set. The line graphs to presents the profit generated and that of the returns over a time period (weekly, monthly, and yearly) thereby showcasing the trends.
Here again, the thematic grey represents a healthy business while the red markings indicate areas where the monthly goals are yet to be reached. 

Customer Details Dashboard - This dashboard offers detailed insights into customer data, covering purchase behaviours, revenue generated per customer and customer segmentation of orders on the basis of income and profession. The line chart highlight the order trends over time and the slicer givens the viewer options to focus o  years individually.